chage float type

+++ fine-tunning/lora-peft-confusion-matrix-competencies-v2.py
@@ -119,7 +119,7 @@ def train_model(configs):
         torch_dtype=torch.bfloat16,
         quantization_config=BitsAndBytesConfig(
             load_in_4bit=True,
-            bnb_4bit_compute_dtype=torch.bfloat16,
+            bnb_4bit_compute_dtype=torch.float32,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type='nf4'
         ),
